+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.8
 
 - Add reader for ASCAT VOD time series data.
 - Add readers for all H-SAF time series products.
+- Automatically detect CGLS SWI-TS time series product date and version.
 
 Version 0.7
 ===========